app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
langchain==0.3.25
langchain-openai==0.3.18
langchain-postgres==0.0.13
langgraph==0.4.1
psycopg[binary]==3.2.0
snowflake-connector-python==3.12.0
azure-cosmos==4.7.0
requests==2.32.0
pydantic==2.9.0
